`testRecursiveBucketDelete` is mysteriously causing `OZONE_FS_HSYNC_ENABLED` to be flipped to `false` in `testAdminCmdListOpenFiles`.

+++ hadoop-ozone/common/src/main/java/org/apache/hadoop/ozone/om/protocol/OzoneManagerProtocol.java
@@ -417,7 +417,7 @@ public interface OzoneManagerProtocol
    * @return ListOpenFilesResult
    * @throws IOException
    */
-  ListOpenFilesResult listOpenFiles(String path, long maxKeys, String contToken)
+  ListOpenFilesResult listOpenFiles(String path, int maxKeys, String contToken)
       throws IOException;
 
   /**
